refactor(movie-detail): extract cast components
- created CastMemberView and CastMoreButtonView
- moved previews to separate files with wrapper views
- added PreviewStyle modifier for consistent layout
- cleaned up PreviewData with longNameMember
